Separating composables and activities in the activities for Chapters 10 and 14
