+++ Chapter10/Activity10.01/app/src/main/java/com/example/tvguide/DetailsActivity.kt
@@ -4,31 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
+import com.example.tvguide.ui.DetailsScreen
 import com.example.tvguide.ui.theme.TVGuideTheme
 
 class DetailsActivity : ComponentActivity() {
@@ -41,96 +17,22 @@ class DetailsActivity : ComponentActivity() {
         const val EXTRA_POSTER = "poster"
     }
 
-    @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
             TVGuideTheme {
-                Scaffold(modifier = Modifier.fillMaxSize(),
-                    topBar = {
-                        TopAppBar(
-                            title = {
-                                Text(
-                                    text = stringResource(id = R.string.app_name),
-                                    color = MaterialTheme.colorScheme.onPrimary
-                                )
-                            },
-                            colors = topAppBarColors(MaterialTheme.colorScheme.primary),
-                            navigationIcon = {
-                                IconButton(onClick = { finish() }) {
-                                    Icon(
-                                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                                        contentDescription = stringResource(id = R.string.back),
-                                        tint = MaterialTheme.colorScheme.onPrimary
-                                    )
-                                }
-                            }
-                        )
-                    }) { innerPadding ->
-                    val extras = intent.extras
-                    val posterPath = extras?.getString(EXTRA_POSTER).orEmpty()
-                    DetailsView(
-                        title = extras?.getString(EXTRA_TITLE).orEmpty(),
-                        release = extras?.getString(EXTRA_RELEASE).orEmpty().take(4),
-                        overview = extras?.getString(EXTRA_OVERVIEW).orEmpty(),
-                        image = "$IMAGE_URL$posterPath",
-                        modifier = Modifier.padding(innerPadding)
-                    )
-                }
+                val extras = intent.extras
+                val posterPath = extras?.getString(EXTRA_POSTER).orEmpty()
+
+                DetailsScreen(
+                    title = extras?.getString(EXTRA_TITLE).orEmpty(),
+                    release = extras?.getString(EXTRA_RELEASE).orEmpty().take(4),
+                    overview = extras?.getString(EXTRA_OVERVIEW).orEmpty(),
+                    image = "$IMAGE_URL$posterPath",
+                    onBackButtonClick = { finish() }
+                )
             }
         }
-    }
-}
-
-@Composable
-fun DetailsView(
-    title: String,
-    release: String,
-    overview: String,
-    image: String,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        AsyncImage(
-            model = image,
-            contentDescription = stringResource(id = R.string.tv_show_poster),
-            contentScale = ContentScale.Fit,
-            modifier = Modifier.height(240.dp)
-        )
-        Text(
-            text = stringResource(id = R.string.tv_show_title, title),
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Text(
-            text = stringResource(id = R.string.tv_show_release, release),
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Text(
-            text = stringResource(id = R.string.tv_show_overview, overview),
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DetailsViewPreview() {
-    TVGuideTheme {
-        DetailsView(
-            title = "TV Show Title",
-            release = "2024-05-06",
-            overview = "TV Show overview line 1\nLine 2",
-            image = "https://image.tmdb.org/t/p/w500/1E5baAaEse26fej7uHcjOgEE2t2.jpg"
-        )
     }
 }
